switch to .hpp file, adjust cmake
