+++ power_grid_model_io_native_c/power_grid_model_io_native_c/CMakeLists.txt
@@ -5,8 +5,6 @@
 # C API library
 add_library(power_grid_model_io_native_c SHARED
   "src/handle.cpp"
-  "src/vnf_pgm_converter.cpp"
-  "src/c_converter_api.cpp"
 )
 
 target_include_directories(power_grid_model_io_native_c PUBLIC
@@ -17,8 +15,6 @@ target_include_directories(power_grid_model_io_native_c PUBLIC
 set(PGM_IO_NATIVE_PUBLIC_HEADERS
   "${CMAKE_CURRENT_SOURCE_DIR}/include/power_grid_model_io_native_c/basics.h"
   "${CMAKE_CURRENT_SOURCE_DIR}/include/power_grid_model_io_native_c/handle.h"
-  "${CMAKE_CURRENT_SOURCE_DIR}/include/power_grid_model_io_native_c/vnf_pgm_converter.h"
-  "${CMAKE_CURRENT_SOURCE_DIR}/include/power_grid_model_io_native_c/c_converter_api.h"
   "${CMAKE_CURRENT_SOURCE_DIR}/include/power_grid_model_io_native_c.h"
 )
