+++ power_grid_model_io_native_c/power_grid_model_io_native_c/include/power_grid_model_io_native_c.h
@@ -18,7 +18,5 @@
 
 #include "power_grid_model_io_native_c/basics.h"
 #include "power_grid_model_io_native_c/handle.h"
-#include "power_grid_model_io_native_c/c_converter_api.h"
-#include "power_grid_model_io_native_c/vnf_pgm_converter.h"
 
 #endif // POWER_GRID_MODEL_IO_NATIVE_C_H
