--- power_grid_model_io_native_c/power_grid_model_io_native_c/include/power_grid_model_io_native_c/vnf_pgm_converter.h
@@ -1,62 +1,0 @@
-// SPDX-FileCopyrightText: Contributors to the Power Grid Model project <powergridmodel@lfenergy.org>
-//
-// SPDX-License-Identifier: MPL-2.0
-
-#include "../../../../../../power-grid-model/power_grid_model_c/power_grid_model_c/src/forward_declarations.hpp"
-
-#pragma once
-#ifndef POWER_GRID_MODEL_IO_NATIVE_C_VNF_PGM_CONVERTER_H
-#define POWER_GRID_MODEL_IO_NATIVE_C_VNF_PGM_CONVERTER_H
-
-#ifdef __cplusplus
-
-class PgmVnfConverter;
-
-extern "C" {
-#endif
-
-PgmVnfConverter* parse_vnf_file_wrapper(PgmVnfConverter* obj);
-
-PGM_ConstDataset* convert_input_wrapper(PgmVnfConverter* obj);
-
-#ifdef __cplusplus
-}
-#endif
-
-
-class PgmVnfConverter {
-    public:
-        PgmVnfConverter();
-        // input data
-        // will be supplied by the user:
-        // ConstDatasetPtr const unique_input_dataset{PGM_create_dataset_const(hl, "input", 0, 1)};
-        // PGM_ConstDataset* input_dataset = unique_input_dataset.get();
-        // PGM_dataset_const_* pgm_input_data;
-
-        // Public member functions
-        void parse_vnf_file();
-        PGM_ConstDataset* convert_input();
-
-    private:
-        // Private attributes
-        char* f_file_buffer;
-        PGM_WritableDataset* deserialized_data; // this type because it is generated by a deserializer type structure
-        // Private setters/getters
-        void set_file_buffer(char* file_buffer);
-        void set_deserialized_data(PGM_WritableDataset* deserialized_data);
-        char* get_file_buffer();
-        PGM_WritableDataset* get_deserialized_data();
-        // Private member functions
-        void convert_node_input();
-        void convert_line_input();
-        void convert_sources_input();
-        void convert_sym_loads_input();
-        void convert_shunts_input();
-        void convert_transformer_input();
-        void convert_sym_gens_input();
-        void convert_links_input();
-};
-
-// void set_core_path(char const* const path) {};
-
-#endif // VNF_PGM_CONVERTER_H
